requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.9
requests==2.32.0
python-dotenv==1.0.0
mcp==1.14.0
atlassian-python-api==4.0.7
python-gitlab==6.3.0
jira==3.10.5
openai==1.107.3
anthropic==0.40.0
google-generativeai==0.8.3
aiohttp==3.10.11
pydantic==2.11.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-ldap==3.4.3
ldap3==2.9.1
python-jwt==4.0.0
redis=6.4.0
PyJWT=2.10.1